services/orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
semantic-kernel>=1.20.0
